refactor(core): split temporal guide types and pipeline temporal tests
Extract pure GuidePlan contracts into guide-plan-types and keep temporal
axis planning / planBasicAxis in temporal-guide (re-exporting types for a
stable public path).

Replace the 1413-line pipeline-temporal monolith with focused suites under
tests/pipeline-temporal (guides, parsers, scales/inference, stats/facets,
annotations/preflight) and point test:temporal-pipeline at the directory.
